+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-orchestration-spring/shardingsphere-jdbc-orchestration-spring-boot-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -89,7 +89,7 @@
     {
       "name": "spring.shardingsphere.datasource",
       "type": "java.util.Map<java.lang.String,java.util.Map<java.lang.String,java.lang.Object>>",
-      "sourceType": "org.apache.shardingsphere.spring.boot.util.DataSourceUtil"
+      "sourceType": "org.apache.shardingsphere.driver.spring.boot.util.DataSourceUtil"
     },
     {
       "sourceType": "org.apache.shardingsphere.sharding.yaml.config.YamlShardingStrategyConfiguration",
