+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-spring-boot-util/src/main/java/org/apache/shardingsphere/driver/spring/boot/datasource/AbstractDbcp2DataSourcePropertiesSetter.java
@@ -15,37 +15,38 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.spring.boot.datasource;
+package org.apache.shardingsphere.driver.spring.boot.datasource;
 
 import lombok.SneakyThrows;
-import org.apache.shardingsphere.spring.boot.util.PropertyUtil;
+import org.apache.shardingsphere.driver.spring.boot.util.PropertyUtil;
 import org.springframework.core.env.Environment;
 
 import javax.sql.DataSource;
 import java.lang.reflect.Method;
 import java.util.Map;
-import java.util.Properties;
 
 /**
- * Hikari datasource properties setter.
+ * DBCP datasource properties setter.
  */
-public final class HikariDataSourcePropertiesSetter implements DataSourcePropertiesSetter {
+public abstract class AbstractDbcp2DataSourcePropertiesSetter implements DataSourcePropertiesSetter {
     
-    @Override
+    /**
+     * Common DBCP2 add custom connection properties.
+     *
+     * @param environment environment variable
+     * @param prefix properties prefix
+     * @param dataSourceName current database name
+     * @param dataSource dataSource instance
+     */
     @SneakyThrows(ReflectiveOperationException.class)
     public void propertiesSet(final Environment environment, final String prefix, final String dataSourceName, final DataSource dataSource) {
-        Properties properties = new Properties();
-        String datasourcePropertiesKey = prefix + dataSourceName.trim() + ".data-source-properties";
-        if (PropertyUtil.containPropertyPrefix(environment, datasourcePropertiesKey)) {
-            Map datasourceProperties = PropertyUtil.handle(environment, datasourcePropertiesKey, Map.class);
-            properties.putAll(datasourceProperties);
-            Method method = dataSource.getClass().getMethod("setDataSourceProperties", Properties.class);
-            method.invoke(dataSource, properties);
+        String datasourcePropertiesPrefix = prefix + dataSourceName.trim() + ".connection-properties";
+        if (PropertyUtil.containPropertyPrefix(environment, datasourcePropertiesPrefix)) {
+            Map datasourceProperties = PropertyUtil.handle(environment, datasourcePropertiesPrefix, Map.class);
+            Method method = dataSource.getClass().getMethod("addConnectionProperty", String.class, String.class);
+            for (Object each : datasourceProperties.keySet()) {
+                method.invoke(dataSource, each, datasourceProperties.get(each));
+            }
         }
-    }
-    
-    @Override
-    public String getType() {
-        return "com.zaxxer.hikari.HikariDataSource";
     }
 }
