+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-spring-boot-util/src/main/java/org/apache/shardingsphere/driver/spring/boot/datasource/DataSourcePropertiesSetter.java
@@ -15,23 +15,31 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.spring.boot.datasource;
+package org.apache.shardingsphere.driver.spring.boot.datasource;
 
-import com.zaxxer.hikari.HikariDataSource;
-import org.junit.Test;
+import org.springframework.core.env.Environment;
 
-import java.util.Optional;
+import javax.sql.DataSource;
 
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-
-public final class DataSourcePropertiesSetterHolderTest {
+/**
+ * Different datasource properties setter.
+ */
+public interface DataSourcePropertiesSetter {
     
-    @Test
-    public void assertGetDataSourcePropertiesSetterByType() {
-        Optional<DataSourcePropertiesSetter> actual = DataSourcePropertiesSetterHolder.getDataSourcePropertiesSetterByType(HikariDataSource.class.getName());
-        assertTrue(actual.isPresent());
-        assertThat(actual.get(), instanceOf(HikariDataSourcePropertiesSetter.class));
-    }
+    /**
+     * Set datasource custom properties.
+     *
+     * @param environment environment variable
+     * @param prefix properties prefix
+     * @param dataSourceName current database name
+     * @param dataSource dataSource instance
+     */
+    void propertiesSet(Environment environment, String prefix, String dataSourceName, DataSource dataSource);
+    
+    /**
+     * Get type name of data source.
+     *
+     * @return type name of data source.
+     */
+    String getType();
 }
