+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-spring-boot-util/src/main/java/org/apache/shardingsphere/driver/spring/boot/util/PropertyUtil.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.spring.boot.util;
+package org.apache.shardingsphere.driver.spring.boot.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
